EOS-18920: Patch-3: module movement for bootstrap logic (#1072)
* creating separate modules for bootstrap logic


* rename file


* r2 test file update for EOS-18920
