+++ test/r2.md
@@ -25,6 +25,7 @@ Table column description
 | Proposed    | [EOS-18743](https://jts.seagate.com/browse/EOS-18917) | | [EOS-18917: Creation of custom grains for enclosure_id,netmask, serial number and multipath](#eos-18917-creation-of-custom-grains-for-enclosure_id,-netmask,-serial-number-and-multipath)                |             |
 | Proposed    | [EOS-17828](https://jts.seagate.com/browse/EOS-17828) | | [EOS-17828: Port adjustments for 3 node](#eos-17828-port-adjustments-post-discussions-for-3-node)                |             |
 | Proposed    | [EOS-17830](https://jts.seagate.com/browse/EOS-17830) | | [EOS-17830: Cluster Bootstrap - Unique Cluster-ID assignment](#cluster-bootstrap-unique-cluster-id-assignment)                |             |
+| Proposed    | [EOS-18920](https://jts.seagate.com/browse/EOS-18920) | | [EOS-18920: Validations for provisioner CLI commands](#eos-18920-validations-for-provisioner-cli-commands)                |             |
 
 
 
@@ -508,3 +509,8 @@ done
 
 ## EOS-17830: Cluster Bootstrap - Unique Cluster-ID assignment
    Steps : Verify if all these scenarios are working fine from this test case doc: [ClusterID-Get-Set Scenarios](https://seagatetechnology.sharepoint.com/:x:/r/sites/eos.devops/Shared%20Documents/Provisioning/QA/ClusterID-Get-Set%20Scenarios.xlsx?d=wb956f3842b634a66b2bed2341786b450&csf=1&web=1&e=QYSbgv)
+
+
+## EOS-18920: Validations for provisioner CLI commands
+   Steps : Please verify if all these scenarios are working fine from this test case doc: [Config-Validations-Scenarios](https://seagatetechnology.sharepoint.com/:x:/r/sites/eos.devops/Shared%20Documents/Provisioning/QA/Config%20Validation%20Scenarios.xlsx?d=w73fbedf43ece422ab7ce6b806d6f8f84&csf=1&web=1&e=otLPMZ)
+
